Refactor: Moved ProfileMenu to SectionMenu folder
